Split default server into own file, rename ServerOps file
