+++ web/src/main/scala/quasar/api/Server.scala
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2014 - 2015 SlamData Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package quasar.api
+
+import quasar.Predef.{Array, Map, None, SuppressWarnings}
+import quasar.config.{ServerConfig, WebConfig, WebConfigLens}
+
+// https://github.com/puffnfresh/wartremover/issues/149
+@SuppressWarnings(Array("org.brianmckenna.wartremover.warts.NonUnitStatements"))
+object Server extends ServerOps(
+  WebConfig,
+  WebConfig(ServerConfig(None), Map()),
+  WebConfigLens(
+    WebConfig.server,
+    WebConfig.mountings,
+    WebConfig.server composeLens ServerConfig.port,
+    ServerConfig.port))
